Adds a quick test to check long requests
There were some reports that requests that exceeded the length of the
time variable would fail. I'm nearly certain that that case is handled
by PyDAP, but I've added a short check here to ensure that such
requests work.
